==> penemue/notebooks/02-Content-Sharing.ipynb <==
import json
import sys
import os
import matplotlib.pyplot as plt
import pandas as pd
import collections

# add penemue to path
sys.path.append(os.path.abspath(os.path.join(os.getcwd(), os.pardir)))
from utils import twiterate
from utils import limiter

%matplotlib inline

def contains_entities(t, search):
    if t['user']['id_str'] in search:
        
        result = []
    
        if 'urls' in t['entities']:
            if len(t['entities']['urls']) > 0:
                result.append('url')
                
        if 'media' in t['entities']:
            if len(t['entities']['media']) > 0:
                for e in t['entities']['media']:
                    if e['type'] == 'photo':
                        result.append('photo')
                
        if 'hashtags' in t['entities']:
            if len(t['entities']['hashtags']) > 0:
                result.append('hashtag')
                
        if 'extended_entities' in t['entities']:
            if len(t['entities']['extended_entities']) > 0:
                for e in t['entities']['extended_entities']:
                    if e['type'] == 'video':
                        result.append('video')
                
        return result

with open('../data/output/journalist_id_strs.json') as f:
    joi = json.load(f)
    joi_callback = lambda t : contains_entities(t, joi)
    joi_entities = collections.Counter([item 
                                        for sublist in twiterate(joi_callback) 
                                        for item in sublist])

with open('../data/output/organisation_id_strs.json') as f:
    ooi = json.load(f)
    ooi_callback = lambda t : contains_entities(t, ooi)
    ooi_entities = collections.Counter([item 
                                        for sublist in twiterate(ooi_callback) 
                                        for item in sublist])

pd.DataFrame(list(joi_entities.values()), 
             list(joi_entities.keys()), 
             ['JOI Tweets'])

pd.DataFrame(list(ooi_entities.values()), 
             list(ooi_entities.keys()), 
             ['OOI Tweets'])
