==> penemue/notebooks/03-Quantities.ipynb <==
import json

with open('../data/output/tweets.json') as tweets:
    print(len([0 for row in tweets]))

with open('../data/output/joi_originals.json') as joi_originals:
    print(len(json.load(joi_originals)))

with open('../data/output/ooi_originals.json') as ooi_originals:
    print(len(json.load(ooi_originals)))

with open('../data/output/joi_retweets.json') as joi_retweets:
    print(len(json.load(joi_retweets)))

with open('../data/output/ooi_retweets.json') as ooi_retweets:
    print(len(json.load(ooi_retweets)))

with open('../data/output/joi_replies.json') as joi_replies:
    print(len(json.load(joi_replies)))

with open('../data/output/ooi_replies.json') as ooi_replies:
    print(len(json.load(ooi_replies)))
