==> penemue/notebooks/03-Unclassified-Users.ipynb <==
import json
import sys
import os

sys.path.append(os.path.abspath(os.path.join(os.getcwd(), os.pardir)))
from utils import twiterate
from utils import lookup

def get_known_user(t):
    """Return only the id_str, description for users.
    
    :param t: A tweet object
    :return:  An id_str (string)
    """
    
    return t['user']['id_str']

def get_unknown_user(t):
    """Extract list of user id_strs from 
    mentions and replies
    
    :param t: A tweet object
    :return:  List of id_strs (strings)
    """
    
    if t['in_reply_to_user_id_str']:
        return [t['in_reply_to_user_id_str']]
    
    if len(t['entities']['user_mentions']) > 0:
        return [m['id_str'] for m in t['entities']['user_mentions']]

known = list(set(twiterate(get_known_user)))

unknown = list(set([item
                    for sublist in twiterate(get_unknown_user)
                    for item in sublist
                    if item not in known]))

print(len(known))
print(len(unknown))

for i in range(0, 200, 100):
    with open('../data/output/unclassified.json', 'w') as f:
        json.dump([{'id_str': user['id_str'], 'description': user['description']}
                   for user in lookup(','.join(unknown[i : i + 100]))], f)
