==> penemue/notebooks/04-User-Extraction.ipynb <==
import sys
import os
import pandas as pd
import numpy as np
import re

sys.path.append(os.path.abspath(os.path.join(os.getcwd(), os.pardir)))
from utils import twiterate

def callback(tweet):
    user = tweet['user']
    description = re.sub(r'\r|\n|\s+', r' ', str(user['description']))
    
    return [user['id_str'], user['screen_name'], description,
            user['followers_count'], tweet['created_at']]

data = twiterate(callback)

df = pd.DataFrame(data, columns=['id_str', 'screen_name', 'description', 
                                 'followers_count', 'date_updated'])

df.shape

df['date_updated'] = pd.to_datetime(df['date_updated'])

df = df.sort_values(['date_updated'])

df = df.dropna()

df = df.drop_duplicates('id_str', keep='last')

df.shape

df.to_json('../data/output/users.json', orient='records')
