==> penemue/notebooks/05-Activity.ipynb <==
import json
import sys
import os.path
import pandas as pd

# add penemue to path
sys.path.append(os.path.abspath(os.path.join(os.getcwd(), os.pardir)))
from utils import twiterate

def get_date(tweet, search):
    if tweet['user']['id_str'] in search:
        return (tweet['id_str'], tweet['created_at'])

journalists = json.load(open('../data/output/journalist_id_strs.json'))

joi_dates = twiterate(lambda t : get_date(t, journalists))

df = pd.DataFrame(joi_dates, columns=['id_str', 'created_at'])

df['created_at'] = pd.to_datetime(df['created_at']).dt.date

df = df.sort_values(['created_at'])

df = df.groupby(['created_at']).count()

df

organisations = json.load(open('../data/output/organisation_id_strs.json'))

ooi_dates = twiterate(lambda t : get_date(t, organisations))

df = pd.DataFrame(ooi_dates, columns=['id_str', 'created_at'])

df['created_at'] = pd.to_datetime(df['created_at']).dt.date

df = df.sort_values(['created_at'])

df = df.groupby(['created_at']).count()

df
