==> penemue/notebooks/05-Tweet-Filtering.ipynb <==
import json
import sys
import os.path
import pandas as pd
import matplotlib.pyplot as plt

%matplotlib inline

# add penemue to path
sys.path.append(os.path.abspath(os.path.join(os.getcwd(), os.pardir)))
from utils import twiterate

def contains_user(tweet, user_ids):
    
    if tweet['in_reply_to_user_id_str'] not in user_ids:
        return False

    for mention in tweet['entities']['user_mentions']:
        if mention['id_str'] not in user_ids:
            return False
        
    return True

with open('../data/output/users.json') as usersfile:
    with open('../data/output/tweets.json') as tweetsfile:
        with open('../data/output/tweets_discussions_within_dataset.json', 'w') as filteredfile:
    
            user_ids = [u['id_str'] for u in json.load(usersfile)]

            for row in tweetsfile:
                tweet = json.loads(row)
                if contains_user(tweet, user_ids):
                    json.dump(tweet, filteredfile)
                    filteredfile.write('\n')
