==> penemue/notebooks/07-Tweet-Sentiment.ipynb <==
import json
import sys
import os
import pandas as pd
from textblob import TextBlob

# add penemue to path
sys.path.append(os.path.abspath(os.path.join(os.getcwd(), os.pardir)))
from utils import twiterate

def get_original(tweet, search):
    if (tweet["user"]["id_str"] in search 
        and tweet["in_reply_to_status_id_str"] is None 
        and "retweeted_status" not in tweet):
            return tweet["text"]

def sentiment(tweets):
    """Get the number of positive, neutral and
    negative tweets in a sample."""
    
    polarity = [TextBlob(text).sentiment.polarity
                for text in tweets]

    positive = len([x 
                    for x in polarity
                         if x > 0])

    neutral = len([x 
                  for x in polarity
                      if x == 0])

    negative = len([x 
                   for x in polarity
                       if x < 0])
    
    return [positive, neutral, negative]

journalists = json.load(open('../data/output/journalist_id_strs.json'))

joi_originals = twiterate(lambda tweet : get_original(tweet, journalists))

pd.DataFrame([sentiment(joi_originals)], 
             columns=['Positive', 'Neutral', 'Negative'])

organisations = json.load(open('../data/output/organisation_id_strs.json'))

ooi_originals = twiterate(lambda tweet : get_original(tweet, organisations))

pd.DataFrame([sentiment(ooi_originals)], 
             columns=['Positive', 'Neutral', 'Negative'])
