==> penemue/notebooks/08-Mention-Network-Extraction.ipynb <==
import json
import sys
import os
import collections
import csv

# add penemue to path
sys.path.append(os.path.abspath(os.path.join(os.getcwd(), os.pardir)))
from utils import twiterate

def get_edge(tweet, search):
    if tweet['user']['id_str'] in search:
        if ('retweeted_status' not in tweet
            and len(tweet["entities"]["user_mentions"]) > 0):
            return [(tweet["user"]["screen_name"], mention["screen_name"])
                    for mention in tweet["entities"]["user_mentions"]]

journalists = json.load(open('../data/output/journalist_id_strs.json'))
organisations = json.load(open('../data/output/organisation_id_strs.json'))
ids = journalists + organisations

edges = [edge 
         for sublist in twiterate(lambda t : get_edge(t, ids)) 
         for edge in sublist]

edges = [[nodes[0], nodes[1], weight] 
         for nodes, weight in collections.Counter(edges).items()]

print(len(edges))

with open("../data/output/mention_edges.csv", 'w') as f:
    w = csv.writer(f)
    w.writerow(["source", "target", "weight"])
    w.writerows(edges)
